+++ src/containerapp/azext_containerapp/custom.py
@@ -470,7 +470,9 @@ def create_containerapp(cmd,
                         service_principal_client_id=None,
                         service_principal_client_secret=None,
                         service_principal_tenant_id=None,
-                        max_inactive_revisions=None):
+                        max_inactive_revisions=None,
+                        runtime=None,
+                        enable_java_metrics=None):
     raw_parameters = locals()
 
     containerapp_create_decorator = ContainerAppPreviewCreateDecorator(
@@ -529,7 +531,9 @@ def update_containerapp_logic(cmd,
                               artifact=None,
                               build_env_vars=None,
                               max_inactive_revisions=None,
-                              force_single_container_updates=False):
+                              force_single_container_updates=False,
+                              runtime=None,
+                              enable_java_metrics=None):
     raw_parameters = locals()
 
     containerapp_update_decorator = ContainerAppPreviewUpdateDecorator(
@@ -580,7 +584,9 @@ def update_containerapp(cmd,
                         source=None,
                         artifact=None,
                         build_env_vars=None,
-                        max_inactive_revisions=None):
+                        max_inactive_revisions=None,
+                        runtime=None,
+                        enable_java_metrics=None):
     _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
 
     return update_containerapp_logic(cmd=cmd,
@@ -616,7 +622,9 @@ def update_containerapp(cmd,
                                      source=source,
                                      artifact=artifact,
                                      build_env_vars=build_env_vars,
-                                     max_inactive_revisions=max_inactive_revisions)
+                                     max_inactive_revisions=max_inactive_revisions,
+                                     runtime=runtime,
+                                     enable_java_metrics=enable_java_metrics)
 
 
 def show_containerapp(cmd, name, resource_group_name, show_secrets=False):
@@ -710,6 +718,7 @@ def create_managed_environment(cmd,
                                certificate_key_vault_url=None,
                                enable_workload_profiles=True,
                                mtls_enabled=None,
+                               p2p_encryption_enabled=None,
                                enable_dedicated_gpu=False,
                                no_wait=False,
                                logs_dynamic_json_columns=False,
@@ -750,6 +759,7 @@ def update_managed_environment(cmd,
                                min_nodes=None,
                                max_nodes=None,
                                mtls_enabled=None,
+                               p2p_encryption_enabled=None,
                                no_wait=False,
                                logs_dynamic_json_columns=None):
     raw_parameters = locals()
@@ -1385,10 +1395,11 @@ def create_containerapps_from_compose(cmd,  # pylint: disable=R0914
     except CLIInternalError:  # pylint: disable=W0702
         logger.info(  # pylint: disable=W1203
             f"Creating the Container Apps managed environment {managed_env_name} under {env_rg} in {location}.")
-        managed_environment = create_containerapps_compose_environment(cmd,
-                                                                       managed_env_name,
-                                                                       env_rg,
-                                                                       tags=tags)
+        managed_environment = create_managed_environment(cmd,
+                                                         name=managed_env_name,
+                                                         resource_group_name=env_rg,
+                                                         tags=tags,
+                                                         location=location)
 
     compose_yaml = load_yaml_file(compose_file_path)
     parsed_compose_file = ComposeFile(compose_yaml)