+++ src/devcenter/azext_devcenter/custom.py
@@ -1205,8 +1205,12 @@ def devcenter_environment_operation_show_logs_by_operation(
     )
     logs_array = []
     for log in logs:
-        logs_string = json.loads(log)
-        logs_array.append(logs_string)
+        if log:
+            try:
+                logs_string = json.loads(log)
+                logs_array.append(logs_string)
+            except json.JSONDecodeError:
+                logs_array.append(log)
     return logs_array