+++ src/serviceconnector-passwordless/azext_serviceconnector_passwordless/custom.py
@@ -12,6 +12,7 @@ def connection_create_ext(cmd, client,  # pylint: disable=too-many-locals,too-ma
                           user_identity_auth_info=None, system_identity_auth_info=None,
                           service_principal_auth_info_secret=None,
                           key_vault_id=None,
+                          app_config_id=None,
                           service_endpoint=None,
                           private_endpoint=None,
                           store_in_connection_string=False,
@@ -20,6 +21,7 @@ def connection_create_ext(cmd, client,  # pylint: disable=too-many-locals,too-ma
                           # Resource.KubernetesCluster
                           cluster=None, scope=None, enable_csi=False,
                           customized_keys=None,
+                          opt_out_list=None,
                           site=None, slot=None,                                  # Resource.WebApp
                           spring=None, app=None, deployment='default',           # Resource.SpringCloud
                           # Resource.*Postgres, Resource.*Sql*
@@ -46,6 +48,8 @@ def connection_create_ext(cmd, client,  # pylint: disable=too-many-locals,too-ma
                                   server, database,
                                   enable_mi_for_db_linker=get_enable_mi_for_db_linker_func(yes),
                                   customized_keys=customized_keys,
+                                  opt_out_list=opt_out_list,
+                                  app_config_id=app_config_id,
                                   **kwargs)